pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hr-ambient-ai"
version = "0.1.0"
description = "HiBob HRIS integration and CLI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-requests>=2.32.0",
]

[project.scripts]
hibob = "hr_ambient_ai.hibob_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hr_ambient_ai"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=hr_ambient_ai --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
